Clean up useGlobusGroups logic, minor styling fixes, improvements to language
