+++ context/app/static/js/components/profile/Summary/Groups.tsx
@@ -1,31 +1,11 @@
 import React from 'react';
-import { readCookie } from 'js/helpers/functions';
 import SectionPaper from 'js/shared-styles/sections/SectionPaper';
 import LabelledSectionText from 'js/shared-styles/sections/LabelledSectionText';
-import { Stack } from '@mui/material';
-import { useAppContext } from '../../Contexts';
-import { useGlobusGroups } from './hooks';
+import Stack from '@mui/material/Stack';
+import { useCurrentUserGlobusGroups } from './hooks';
 
 export function ProfileGroups() {
-  const { isAuthenticated } = useAppContext();
-  const groupsCookie = readCookie('user_groups');
-  const { data: allGlobusGroups = [] } = useGlobusGroups();
-  if (!groupsCookie || !isAuthenticated) {
-    return null;
-  }
-
-  const groups = groupsCookie.split('|');
-
-  const currentUserGroups = allGlobusGroups
-    ?.filter((group) => groups.includes(group.name))
-    .map((g) => ({
-      key: g.name,
-      name: g.displayname,
-      description: g.description,
-    }));
-
-  const nonHubmapGroups = groups.filter((group) => !currentUserGroups.find((g) => g.key === group));
-
+  const currentUserGroups = useCurrentUserGlobusGroups();
   return (
     <SectionPaper>
       <Stack direction="column" spacing={1}>
@@ -34,9 +14,6 @@ export function ProfileGroups() {
             {description}
           </LabelledSectionText>
         ))}
-        {nonHubmapGroups.length > 0 && (
-          <LabelledSectionText label="Other Groups">{nonHubmapGroups.join(', ')}</LabelledSectionText>
-        )}
       </Stack>
     </SectionPaper>
   );
